Update zzz.cs
Converting the offset to an INT so that will break things if we get the files large enough.

+++ zzz/zzz.cs
@@ -358,7 +358,7 @@ namespace ZzzFile
                     {
                         foreach (FileData i in data)
                         {
-                            _br.BaseStream.Seek((int)i.Offset, SeekOrigin.Begin);
+                            _br.BaseStream.Seek(i.Offset, SeekOrigin.Begin);
                             Console.WriteLine($"Writing {i.Filename} {i.Size} bytes");
                             merged.bw.Write(_br.ReadBytes(i.Size));
                         }
